requirements.txt
Flask==2.0.1
requests==2.26.0
pyaudio==0.2.11
deepgram-sdk==1.0.0
tkinter==0.1.0